refactoring codes
